+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # 6.0.0
 
 ## Breaking Changes
+
 - Vue 2 support has been dropped.
 - The CommonJS build of coalesce-vue has been dropped.
 
 ## Major Changes
+
 - All endpoints with bodies are now send as JSON instead of form data, with the exception of those endpoints with file parameters and only scalar parameters (no complex parameters) beyond the file parameters.
 
+## Features
+
+- Added `$search`, `$filter`, and `$orderBy` shorthands to ListViewModel.
 
 # 5.3.1